[sw-sysemu] Rationalize use of IO_SHIRE constants
Create helper functions to change between processor shire ids and
actual mapping inside sysemu.

Note:

The better, long term solution to this is to create two
different types, index and id, and be able to conver them, to
allow the problem of id and indexes to be solved by type safety.

This is future work because at the moment both indices and ids are
checked in some functions.